requirements.txt
colorama==0.4.4
colorlog==4.7.2
git-filter-repo==2.38.0
jupyter==1.0.0
jupyterlab==4.0.8
matplotlib==3.7.2
numpy==1.23.5
pandas==2.0.3
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
tensorboard==2.12.1
torch==2.1.0
torchmetrics==1.1.2
tqdm==4.65.0
